backend/requirements.txt
fastapi
uvicorn
openai
supabase
twilio
python-dotenv
pydantic
python-multipart
httpx
